refactor(cmdparse): reanme shellcompletion -> completion
