+++ pigit/cmdparse/completion/base.py
