+++ pigit/cmdparse/completion/bash.py
