+++ tests/test_cmdparser.py
@@ -4,8 +4,8 @@ from .utils import analyze_it
 from .conftest import _PIGIT_PATH
 
 from pigit.cmdparse.parser import command, Parser
-from pigit.cmdparse.shellcompletion.base import ShellCompletion
-from pigit.cmdparse.shellcompletion import (
+from pigit.cmdparse.completion.base import ShellCompletion
+from pigit.cmdparse.completion import (
     ZshCompletion,
     BashCompletion,
     FishCompletion,
